# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Gym Management API",
    "description": "REST API for managing gym members, trainers, classes, and memberships.",
    "version": "2.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/status": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}